fix: register 0010 migration in journal, keep 0006 immutable, set backup id default, and fix patch diff semantics

+++ apps/main-node/migrations/0006_session_memory_freeze.sql
@@ -12,4 +12,4 @@
 -- drizzle schema (CF never gets the column), so a snapshot containing it
 -- would make the next drizzle-kit generate emit a DROP COLUMN.
 
-ALTER TABLE "sessions" ADD COLUMN "memory_frozen_at" bigint;
+ALTER TABLE "sessions" ADD COLUMN "memory_frozen_at" integer;
